requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==20.*
google-genai